requirements.txt
pascal_voc_writer
streamlit==1.10.0
streamlit-img-label==0.1.1
streamlit-option-menu==0.3.2
extra_streamlit_components
torch==1.7.1
torchvision==0.8.2
PyYAML==6.0
matplotlib>=3.5.2
numpy==1.22.4
Pillow==9.1.1
requests==2.28.0
scipy>=1.8.1
tqdm>=4.64.0
pandas==1.4.2
seaborn==0.11.2
protobuf==3.19.4
opencv-python==4.6.0.66
celery==5.2.7
redis==4.3.4
#eventlet
#kombu